2008-12-15	Joel Sherrill <joel.sherrill@oarcorp.com>
	* malloctest/init.c: Change sizes of heap/region and allocated objects
	in heap to intptr_t so they can be larger than a single allocatable
	object (e.g. size_t).

+++ testsuites/libtests/malloctest/init.c
@@ -119,7 +119,7 @@ void test_heap_init()
 void test_heap_cases_1()
 {
   void     *p1, *p2, *p3, *p4;
-  uint32_t  u1, u2;
+  intptr_t  u1, u2;
   Heap_Resize_status rsc;
 
   /*
